refactor(file-context): remove file-context
