+++ libs/domain/lib/context.ts
@@ -1,6 +1,6 @@
-import { FileStore } from '@pl-oss/file-context';
 import { EventStore } from './event-store';
 import { FeatureService } from './feature-service';
+import { FileStore } from './file-store';
 import { GraphQLService } from './graphql.service';
 import { LoggingService } from './logging-service';
