--- libs/file-context/README.md
@@ -1,1 +1,0 @@
-# File Context
