--- libs/file-context/index.ts
@@ -1,4 +1,0 @@
-export * from './lib/file.entity';
-export * from './lib/file-store';
-export * from './lib/id-should-be-unique.exception';
-export * from './lib/mongo-db.file-store';
