--- libs/file-context/jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  displayName: 'file-context',
-  preset: '../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]s?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'js'],
-  coverageDirectory: '../../coverage/libs/file-context',
-};
