--- libs/file-context/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@pl-oss/file-context",
-  "version": "0.0.56",
-  "license": "MIT"
-}
