--- libs/file-context/tsconfig.spec.json
@@ -1,16 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "module": "commonjs",
-    "types": [
-      "jest",
-      "node"
-    ]
-  },
-  "include": [
-    "**/*.spec.ts",
-    "**/*.spec.js",
-    "**/*.d.ts"
-  ]
-}
