+++ nx.json
@@ -30,9 +30,6 @@
     "domain": {
       "tags": []
     },
-    "file-context": {
-      "tags": []
-    },
     "vue": {
       "tags": []
     }
